+++ src/counter_analysis_toolkit/driver.h
@@ -29,6 +29,6 @@ int perm(int n, int k);
 int comb(int n, int k);
 void testbench(char** allevts, int cmbtotal, hw_desc_t *hw_desc, cat_params_t params, int myid, int nprocs);
 void print_usage();
-static int parse_line(FILE *input, char **key, int *value);
+static int parse_line(FILE *input, char **key, long long *value);
 static void read_conf_file(char *conf_file, hw_desc_t *hw_desc);
 static hw_desc_t *obtain_hardware_description(char *conf_file_name);
